Use #send_data when handling thumbnails (for now)

+++ app/controllers/documents_controller.rb
@@ -67,13 +67,22 @@ class DocumentsController < ApplicationController
 
   def generate_thumbnail
     @document = Document.find(params[:id])
+
+    # Create thumbnail file in public assets directory
     path = File.join(Rails.root, "public", request.path)
-    File.open(path, "wb") do |fd|
-      @document.thumbnail_file.each do |chunk|
-        fd.write(chunk)
+    if not File.exists?(path)
+      File.open(path, "wb") do |fd|
+        @document.thumbnail_file.each do |chunk|
+          fd.write(chunk)
+        end
       end
     end
-    redirect_to request.path
+
+    # FIXME For now, #send_data here, ideally this should be handled by the
+    # assets server (e.g. nginx).
+    File.open(path, "rb") do |fd|
+      send_data fd, filename: request.path
+    end
   rescue Mongoid::Errors::DocumentNotFound
     render :text => nil, :status => 404
   end
